requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
aiohttp==3.9.1
pytest==7.4.3 